+++ mayan/apps/cabinets/tests/test_wizard_steps.py
@@ -40,7 +40,7 @@ class CabinetDocumentUploadTestCase(
 
         self._clear_events()
 
-        response = self._request_test_source_document_upload_view_with_cabinets()
+        response = self._request_test_source_document_upload_post_view_with_cabinets()
 
         self.assertEqual(response.status_code, 302)
         self.assertTrue(