+++ mayan/apps/metadata/tests/test_wizard_steps.py
@@ -43,7 +43,7 @@ class DocumentUploadMetadataTestCase(
 
         self._clear_events()
 
-        response = self._request_test_source_document_upload_view_with_metadata(
+        response = self._request_test_source_document_upload_post_view_with_metadata(
             metadata_value=TEST_METADATA_VALUE_UNICODE
         )
         self.assertEqual(response.status_code, 302)
@@ -110,7 +110,7 @@ class DocumentUploadMetadataTestCase(
 
         self._clear_events()
 
-        response = self._request_test_source_document_upload_view_with_metadata(
+        response = self._request_test_source_document_upload_post_view_with_metadata(
             metadata_value=TEST_METADATA_VALUE_WITH_AMPERSAND
         )