+++ mayan/apps/source_sane_scanners/tests/test_views.py
@@ -534,7 +534,7 @@ class SANEScannerSourceBackendActionDocumentUploadViewTestCase(
 
         self._clear_events()
 
-        response = self._request_test_source_document_upload_view()
+        response = self._request_test_source_document_upload_post_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertEqual(Document.objects.count(), document_count)
@@ -554,7 +554,7 @@ class SANEScannerSourceBackendActionDocumentUploadViewTestCase(
 
         self._clear_events()
 
-        response = self._request_test_source_document_upload_view()
+        response = self._request_test_source_document_upload_post_view()
         self.assertEqual(response.status_code, 302)
 
         self.assertEqual(Document.objects.count(), document_count)
@@ -573,7 +573,7 @@ class SANEScannerSourceBackendActionDocumentUploadViewTestCase(
 
         self._clear_events()
 
-        response = self._request_test_source_document_upload_view()
+        response = self._request_test_source_document_upload_post_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertEqual(Document.objects.count(), document_count)
@@ -596,7 +596,7 @@ class SANEScannerSourceBackendActionDocumentUploadViewTestCase(
 
         self._clear_events()
 
-        response = self._request_test_source_document_upload_view()
+        response = self._request_test_source_document_upload_post_view()
         self.assertEqual(response.status_code, 302)
 
         self.assertEqual(Document.objects.count(), document_count + 1)
@@ -655,7 +655,7 @@ class SANEScannerSourceBackendActionDocumentUploadViewTestCase(
 
         self._clear_events()
 
-        response = self._request_test_source_document_upload_view(
+        response = self._request_test_source_document_upload_post_view(
             extra_data={'document-language': 'deu'}
         )
         self.assertEqual(response.status_code, 302)